requirements.txt
demucs
ffmpeg-python
transformers
TTS